Fulfil Domain class functionality within Variable
