+++ probayes/rv.py
@@ -5,7 +5,7 @@ import collections
 import warnings
 import numpy as np
 import scipy.stats
-from probayes.domain import Domain, DEFAULT_VNAME
+from probayes.variable import Variable, DEFAULT_VNAME
 from probayes.prob import Prob, is_scipy_stats_cont
 from probayes.dist import Dist
 from probayes.vtypes import eval_vtype, uniform, VTYPES, isscalar, \
@@ -23,9 +23,9 @@ associated probability distribution function (prob).
 """
 
 #-------------------------------------------------------------------------------
-class RV (Domain, Prob):
-  """ A random variable is a domain with a defined probability function.
-  It therefore inherits from classes Domain and and Prob. Each instance therefore 
+class RV (Variable, Prob):
+  """ A random variable is a variable with a defined probability function.
+  It therefore inherits from classes Variable and and Prob. Each instance therefore 
   requires a name, a variable set, and probability function. Additionally RV
   supports transitional probabilities the cdf/icdf equivalents specified using
   RV.set_tran() and RV.set_tfun() equivalents, and accessed using RV.step().
@@ -55,7 +55,7 @@ class RV (Domain, Prob):
                      vtype=None,
                      *args,
                      **kwds):
-    """ Initialises a random variable combining Domain and Prob initialisation
+    """ Initialises a random variable combining Variable and Prob initialisation
     except invertible monotonic must be specified separately using set_mfun().
 
     :param name: Name of the domain - string as valid identifier.
@@ -65,7 +65,7 @@ class RV (Domain, Prob):
     :param *kwds: optional keywords to pass onto symbol representation.
     """
 
-    Domain.__init__(self, name, vset, vtype, *args, **kwds)
+    Variable.__init__(self, name, vset, vtype, *args, **kwds)
     self.set_prob()
 
 #-------------------------------------------------------------------------------
@@ -241,7 +241,7 @@ class RV (Domain, Prob):
     :param values: None, set of a single integer, array, or scalar.
     :param use_pfun: boolean flag to make use of pfun if previously set.
 
-    :return: a NumPy array of the values (see Domain.eval_vals()):
+    :return: a NumPy array of the values (see Variable.eval_vals()):
     """
     use_pfun = use_pfun and self._pfun is not None and isunitsetint(values)
     if not use_pfun:
